refactor: organize files better
