+++ src/bindings/utils/NoBinaryFoundError.ts
