+++ src/bindings/utils/getExampleUsageCodeOfGetLlama.ts
