+++ src/bindings/utils/logBinaryUsageExampleToConsole.ts
