+++ src/cli/commands/DownloadCommand.ts
@@ -6,20 +6,20 @@ import {
     defaultLlamaCppCudaSupport, defaultLlamaCppGitHubRepo, defaultLlamaCppMetalSupport, defaultLlamaCppRelease, isCI, llamaCppDirectory,
     llamaCppDirectoryInfoFilePath
 } from "../../config.js";
-import {compileLlamaCpp} from "../../utils/compileLLamaCpp.js";
+import {compileLlamaCpp} from "../../bindings/utils/compileLLamaCpp.js";
 import withOra from "../../utils/withOra.js";
 import {clearTempFolder} from "../../utils/clearTempFolder.js";
-import {setBinariesGithubRelease} from "../../utils/binariesGithubRelease.js";
+import {setBinariesGithubRelease} from "../../bindings/utils/binariesGithubRelease.js";
 import {downloadCmakeIfNeeded} from "../../utils/cmake.js";
 import withStatusLogs from "../../utils/withStatusLogs.js";
 import {getIsInDocumentationMode} from "../../state.js";
 import {getGitBundlePathForRelease, unshallowAndSquashCurrentRepoAndSaveItAsReleaseBundle} from "../../utils/gitReleaseBundles.js";
-import {cloneLlamaCppRepo} from "../../utils/cloneLlamaCppRepo.js";
-import {getPlatform} from "../../llamaBin/utils/getPlatform.js";
-import {resolveCustomCmakeOptions} from "../../llamaBin/utils/resolveCustomCmakeOptions.js";
-import {logBinaryUsageExampleToConsole} from "../../llamaBin/utils/logBinaryUsageExampleToConsole.js";
+import {cloneLlamaCppRepo} from "../../bindings/utils/cloneLlamaCppRepo.js";
+import {getPlatform} from "../../bindings/utils/getPlatform.js";
+import {resolveCustomCmakeOptions} from "../../bindings/utils/resolveCustomCmakeOptions.js";
+import {logBinaryUsageExampleToConsole} from "../../bindings/utils/logBinaryUsageExampleToConsole.js";
 import {resolveGithubRelease} from "../../utils/resolveGithubRelease.js";
-import {BuildOptions} from "../../llamaBin/types.js";
+import {BuildOptions} from "../../bindings/types.js";
 
 type DownloadCommandArgs = {
     repo?: string,
