+++ src/evaluator/LlamaJsonSchemaGrammar.ts
@@ -2,7 +2,7 @@ import {GbnfJsonSchema, GbnfJsonSchemaToType} from "../utils/gbnfJson/types.js";
 import {getGbnfGrammarForGbnfJsonSchema} from "../utils/getGbnfGrammarForGbnfJsonSchema.js";
 import {validateObjectAgainstGbnfSchema} from "../utils/gbnfJson/utils/validateObjectAgainstGbnfSchema.js";
 import {LlamaText} from "../utils/LlamaText.js";
-import {Llama} from "../llamaBin/Llama.js";
+import {Llama} from "../bindings/Llama.js";
 import {LlamaGrammar} from "./LlamaGrammar.js";
 
 export class LlamaJsonSchemaGrammar<const T extends Readonly<GbnfJsonSchema>> extends LlamaGrammar {
