+++ src/index.ts
@@ -1,31 +1,31 @@
 import {DisposedError} from "lifecycle-utils";
-import {Llama} from "./llamaBin/Llama.js";
-import {getLlama, LlamaOptions} from "./llamaBin/getLlama.js";
-import {NoBinaryFoundError} from "./llamaBin/utils/NoBinaryFoundError.js";
-import {LlamaLogLevel} from "./llamaBin/types.js";
-import {LlamaModel, LlamaModelInfillTokens, type LlamaModelOptions, LlamaModelTokens} from "./llamaEvaluator/LlamaModel.js";
-import {LlamaGrammar, type LlamaGrammarOptions} from "./llamaEvaluator/LlamaGrammar.js";
-import {LlamaJsonSchemaGrammar} from "./llamaEvaluator/LlamaJsonSchemaGrammar.js";
+import {Llama} from "./bindings/Llama.js";
+import {getLlama, LlamaOptions} from "./bindings/getLlama.js";
+import {NoBinaryFoundError} from "./bindings/utils/NoBinaryFoundError.js";
+import {LlamaLogLevel} from "./bindings/types.js";
+import {LlamaModel, LlamaModelInfillTokens, type LlamaModelOptions, LlamaModelTokens} from "./evaluator/LlamaModel.js";
+import {LlamaGrammar, type LlamaGrammarOptions} from "./evaluator/LlamaGrammar.js";
+import {LlamaJsonSchemaGrammar} from "./evaluator/LlamaJsonSchemaGrammar.js";
 import {LlamaJsonSchemaValidationError} from "./utils/gbnfJson/utils/validateObjectAgainstGbnfSchema.js";
-import {LlamaGrammarEvaluationState, LlamaGrammarEvaluationStateOptions} from "./llamaEvaluator/LlamaGrammarEvaluationState.js";
-import {LlamaContext, LlamaContextSequence} from "./llamaEvaluator/LlamaContext/LlamaContext.js";
+import {LlamaGrammarEvaluationState, LlamaGrammarEvaluationStateOptions} from "./evaluator/LlamaGrammarEvaluationState.js";
+import {LlamaContext, LlamaContextSequence} from "./evaluator/LlamaContext/LlamaContext.js";
 import {
     LlamaEmbeddingContext, type LlamaEmbeddingContextOptions, LlamaEmbedding, type LlamaEmbeddingJSON
-} from "./llamaEvaluator/LlamaEmbeddingContext.js";
+} from "./evaluator/LlamaEmbeddingContext.js";
 import {
     type LlamaContextOptions, type BatchingOptions, type LlamaContextSequenceRepeatPenalty, type CustomBatchingDispatchSchedule,
     type CustomBatchingPrioritizeStrategy, type BatchItem, type PrioritizedBatchItem, type ContextShiftOptions,
     type ContextTokensDeleteRange, type EvaluationPriority
-} from "./llamaEvaluator/LlamaContext/types.js";
+} from "./evaluator/LlamaContext/types.js";
 import {
     LlamaChatSession, type LlamaChatSessionOptions, type LlamaChatSessionContextShiftOptions,
     type LLamaChatPromptOptions, type LlamaChatSessionRepeatPenalty
-} from "./llamaEvaluator/LlamaChatSession/LlamaChatSession.js";
-import {defineChatSessionFunction} from "./llamaEvaluator/LlamaChatSession/utils/defineChatSessionFunction.js";
+} from "./evaluator/LlamaChatSession/LlamaChatSession.js";
+import {defineChatSessionFunction} from "./evaluator/LlamaChatSession/utils/defineChatSessionFunction.js";
 import {
     LlamaChat, type LlamaChatOptions, type LLamaChatGenerateResponseOptions, type LLamaChatContextShiftOptions,
     type LLamaChatRepeatPenalty, type LlamaChatResponse, type LlamaChatResponseFunctionCall
-} from "./llamaEvaluator/LlamaChat/LlamaChat.js";
+} from "./evaluator/LlamaChat/LlamaChat.js";
 import {AbortError} from "./AbortError.js";
 import {ChatWrapper, type ChatWrapperSettings} from "./ChatWrapper.js";
 import {EmptyChatWrapper} from "./chatWrappers/EmptyChatWrapper.js";
